requirements.txt
absl-py==0.11.0
alabaster==0.7.12
argon2-cffi==20.1.0
astor==0.8.1
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorlog==4.7.2
cvxpy==1.1.10
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
docutils==0.16
ecos==2.0.7.post1
entrypoints==0.3
Flask==1.1.2
gast==0.4.0
google-pasta==0.2.0
grpcio==1.35.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imagesize==1.2.0
imgaug==0.4.0
importlib-metadata==2.1.1
ipykernel==5.4.3
ipyparallel==6.3.0
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.6.3
Keras==2.0.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
mask-rcnn==2.1
matplotlib==3.0.3
mistune==0.8.4
mss==6.1.0
nbconvert==5.6.1
nbformat==5.1.2
networkx==2.4
nose==1.3.7
notebook==6.2.0
numpy==1.18.5
opencv-contrib-python==4.4.0.42
opencv-python==4.4.0.42
osqp==0.6.2.post0
packaging==20.9
pafy==0.5.5
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.9.0
prompt-toolkit==2.0.10
protobuf==3.14.0
pycocotools==2.0.2
pycparser==2.20
Pygments==2.8.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==20.0.0
qdldl==0.1.5.post0
qtconsole==5.0.0
QtPy==1.9.0
requests==2.25.1
scikit-image==0.15.0
scipy==1.4.1
scs==2.1.2
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
snowballstemmer==2.1.0
Sphinx==3.5.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.1
tqdm==4.56.2
traitlets==4.3.3
urllib3==1.26.3
vidgear==0.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
youtube-dl==2021.2.10
zipp==1.2.0
